+++ src/features/auth/api/auth.ts
@@ -1,9 +1,4 @@
-import type {
-  LoginPayload,
-  LoginResponse,
-  UserInfo,
-  ShareCodeOwnerResponse,
-} from '@/features/auth/types/auth'
+import type { LoginPayload, LoginResponse, UserInfo } from '@/features/auth'
 import { api } from '@/shared/api/httpClient'
 
 // 회원 로그인 (카카오)
@@ -24,9 +19,4 @@ export const deleteAccount = () => {
 // 채팅 내 이름/아이디 확인
 export const getUserInfo = () => {
   return api.get<UserInfo>('/chat/user')
-}
-
-// shareCode 유효성 및 본인 여부 확인
-export const getShareCodeOwner = (shareCode: string) => {
-  return api.get<ShareCodeOwnerResponse>(`/main/user/profile/${shareCode}/owner`)
 }
