+++ src/pages/home/ui/FirstSection.tsx
@@ -7,7 +7,7 @@ import { Logo, RightArrow, Search } from '@/assets/icons'
 import { HomeCharacter } from '@/assets/images'
 import { usePlaylistDetails, useShufflePlaylists } from '@/entities/playlist'
 import { useMyCdList } from '@/entities/playlist/model/useMyCd'
-import { useAuthStore } from '@/features/auth/store/authStore'
+import { useAuthStore } from '@/features/auth'
 import { FeedbackIcon } from '@/pages/feedback/ui'
 import { BUTTON_TEXT, TITLE_TEXT } from '@/pages/home/config/messages'
 import { HomeCarousel } from '@/pages/home/ui'
