+++ src/pages/login/index.tsx
@@ -6,8 +6,7 @@ import styled from 'styled-components'
 import { useToast } from '@/app/providers'
 import { Kakao } from '@/assets/icons'
 import { LoginBg as LoginBgImg, LoginContent as LoginContentImg } from '@/assets/images'
-import { startKakaoLogin } from '@/features/auth/lib/kakaoLogin'
-import { useAuthStore } from '@/features/auth/store/authStore'
+import { startKakaoLogin, useAuthStore } from '@/features/auth'
 import { flexColCenter, flexRowCenter } from '@/shared/styles/mixins'
 
 const LoginPage = () => {
