+++ src/pages/mypage/tracklist/index.tsx
@@ -7,7 +7,7 @@ import { useToast } from '@/app/providers'
 import { LeftArrow, StartBlack, Trash, Share, Eye, EyeSlash } from '@/assets/icons'
 import { useMyCdActions } from '@/entities/playlist/model/useMyCd'
 import { usePlaylistDetail } from '@/entities/playlist/model/usePlaylists'
-import { useAuthStore } from '@/features/auth/store/authStore'
+import { useAuthStore } from '@/features/auth'
 import { ShareButton } from '@/features/share'
 import { MUSIC_GENRES } from '@/shared/config/musicGenres'
 import { flexRowCenter } from '@/shared/styles/mixins'
